Local Changes

+++ app/src/main/java/com/octalabs/challetapp/models/ModelBookingHistory/BookingHistoryDetails.java
@@ -13,7 +13,7 @@ public class BookingHistoryDetails {
     private List<String> picture = null;
     @SerializedName("rating")
     @Expose
-    private Integer rating;
+    private float rating;
     @SerializedName("pricePerNight")
     @Expose
     private Integer pricePerNight;
@@ -78,11 +78,11 @@ public class BookingHistoryDetails {
         this.picture = picture;
     }
 
-    public Integer getRating() {
+    public float getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
